Adds initial support for Joomla 1.6 and adds ticket type support
The Joomla component has now been split into joomla_1.5 and
joomla_1.6 paths to allow for Joomla 1.6 support. This is for
GitHub issue #5.

Support is also provided for the OTRS ticket type field, which is
enabled in OTRS by setting the Ticket::Type sysconfig value.
An additional ticket type field is shown in the ticket submission
form if Ticket::Type is enabled. See GitHub issue #14.