Use less common form names (GH #22)
The form field names for the submission and reply text fields
have now been changed from message and replyText to otrsMessage
and otrsReplyText respectively.

This change reduces the chances of a naming collision between
the gateway templates and other Joomla templates.

Note that where people have overridden the submission or reply
templates, the overridden templates will need to be updated to
reflect this change.

Affects the templates:

 * views/ticket/tmpl/reply.php
 * views/ticket/tmpl/submit.php