New Backend Setting for Systemmail
If Sender is not Systemmail CSS Class "otrs-reply-self" is used
everytime. If Systemmail is empty, CSS Class is used in case the sender
is Joomlauser Mailaddress.

Its usefull for use with company Tickets. If more then one joomla users
are watching same tickets, only the supporter Mails are with different
css Class.